Now with a new IP IP! (Used in Discrepancies IP)
new is a very simple interpolated prompt (IP) to quickly generate new
IPs.

Then, I used it to create a nice tool to conduct a gap analysis between a resume and a job posting. Here's how that was done, as an example:

- I went to /new.html?PROMPTS=1%0A2%0A3%0A4
- and filled in each template until it felt right.
- Then, I manually saved them into discrepancies.* files.

From there, I iterated to get the version committed before you.